changes commit

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
-    public double calculate(double meters, int kg) {
-        double bmi = kg / (meters*meters);
+    public int calculate(double meters, int kg) {
+        int bmi = (int) (kg / (meters*meters));
         return bmi;
     }
 }
