+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BmiService service = new BmiService();
         double meters = 1.87;
         int kg = 98;
-        double bmi = service.calculate(meters, kg);
-        System.out.println((int)bmi);
+        int bmi = service.calculate(meters, kg);
+        System.out.println(bmi);
     }
 }
